+++ src/entities/IAddSymptomFormStateKey.ts
@@ -1,0 +1,3 @@
+import { IAddSymptomFormState } from "./IAddSymptomFormState";
+
+export type IAddSymptomFormStateKey = keyof IAddSymptomFormState;
