+++ src/entities/IAddSymptomFormStateKeyValue.ts
@@ -1,0 +1,3 @@
+import { IOption } from "./IOption";
+
+export type IAddSymptomFormStateKeyValue = IOption | Date | string;
