--- src/entities/IAddSymptomStateKey.ts
@@ -1,3 +1,0 @@
-import { IAddSymptomState } from "./IAddSymptomFormState";
-
-export type IAddSymptomStateKey = keyof IAddSymptomState;
