--- src/entities/IAddSymptomStateKeyValue.ts
@@ -1,4 +1,0 @@
-import { IOption } from "./IOption";
-import { ISymptom } from "./ISymptom";
-
-export type IAddSymptomStateKeyValue = IOption | Date | string;
