+++ src/screens/ScreenAppAddSymptomProgress/context/index.tsx
@@ -2,8 +2,8 @@ import { createContext, useContext, useState } from "react";
 import { IProviderProps } from "../../../entities/IProviderProps";
 import { useCustomToast } from "../../../hooks/useCustomToast";
 import { INITAL_OPTION } from "../../../data/defaultValues";
-import { IAddSymptomStateKey } from "../../../entities/IAddSymptomStateKey";
-import { IAddSymptomStateKeyValue } from "../../../entities/IAddSymptomStateKeyValue";
+import { IAddSymptomStateKey } from "../../../entities/IAddSymptomFormStateKey";
+import { IAddSymptomStateKeyValue } from "../../../entities/IAddSymptomFormStateKeyValue";
 import { IOption } from "../../../entities/IOption";
 import { useSeverityRatings } from "../../../hooks/useSeverityRatings";
 import { services } from "../../../services";
