+++ src/utils/findOption/index.ts
@@ -4,7 +4,9 @@ import { IOption } from "../../entities/IOption";
 export const findOption: IFindOptionUtil = (items, key, value) => {
   if (!value) return INITAL_OPTION;
 
-  return items?.find((item) => item[key] === value);
+  return items?.find((item) => {
+    return item[key] == value;
+  });
 };
 
 interface IFindOptionUtil {
